+++ src/lib/ColorInput.svelte
@@ -36,9 +36,7 @@
   }
 
   let parent: HTMLElement
-  function blur(e: FocusEvent) {
-    console.log('blur', e.relatedTarget, e.currentTarget, document.activeElement)
-
+  function focusout(e: FocusEvent) {
     if (e.relatedTarget === null) {
       isOpen = false
     } else if (e.relatedTarget instanceof HTMLElement) {
@@ -60,7 +58,13 @@
   }
 </script>
 
-<div bind:this={parent} class="input {classes}" on:mousedown={open} tabindex="-1">
+<div
+  bind:this={parent}
+  class="input {classes}"
+  on:mousedown={open}
+  on:focusout={focusout}
+  tabindex="-1"
+>
   <div class="color-frame">
     <div class="color-frame-color" style:background-color={color.toHex8String()} />
   </div>
@@ -72,7 +76,6 @@
       bind:value={text}
       on:input={onInput}
       on:focus={open}
-      on:blur={blur}
     />
     <span class:show={!isOpen} class="title">{title}</span>
   </div>