+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Next
-- Fix picker not hiding
 - Add `onInput` event
+- Fix picker not hiding
 
 ## 0.4.0 - 2022 Aug 5
 - Add alpha support